why the hell was this never implemented? you should not be able to delete users with readonly sets

+++ lib/_WV16/User.php
@@ -271,6 +271,8 @@ class _WV16_User extends WV_Object implements WV16_User {
 	}
 
 	public function addGroup($group) {
+		if ($this->isReadOnly()) return false;
+
 		$service = new _WV16_Service_User();
 
 		if ($service->addToGroup($this, $group)) {
@@ -282,6 +284,8 @@ class _WV16_User extends WV_Object implements WV16_User {
 	}
 
 	public function removeGroup($group) {
+		if ($this->isReadOnly()) return false;
+
 		$service = new _WV16_Service_User();
 		$service->removeFromGroup($this, $group);
 
@@ -295,13 +299,17 @@ class _WV16_User extends WV_Object implements WV16_User {
 	}
 
 	public function removeAllGroups() {
+		if ($this->isReadOnly()) return false;
 		$service = new _WV16_Service_User();
 		return $service->removeFromAllGroups($this);
 	}
 
 	public function setConfirmationCode($code = null) {
+		if ($this->isReadOnly()) return false;
+
 		$code = $code === null ? WV16_Users::generateConfirmationCode($this->login) : substr($code, 0, 20);
 		$this->confirmationCode = $code;
+
 		return $code;
 	}
 
@@ -310,6 +318,8 @@ class _WV16_User extends WV_Object implements WV16_User {
 
 
 	public function setConfirmed($isConfirmed = true, $confirmationCode = null) {
+		if ($this->isReadOnly()) return false;
+
 		// Auf Wunsch kann auch diese Methode die Überprüfung auf den
 		// Bestätigungscode selbst durchführen.
 
@@ -322,6 +332,7 @@ class _WV16_User extends WV_Object implements WV16_User {
 	}
 
 	public function setActivated($isActivated = true) {
+		if ($this->isReadOnly()) return false;
 		$this->activated = (boolean) $isActivated;
 	}
 
@@ -426,5 +437,9 @@ class _WV16_User extends WV_Object implements WV16_User {
 
 	public function _setEverActivated() {
 		$this->wasActivated = true;
+	}
+
+	public function _setDeleted() {
+		$this->deleted = true;
 	}
 }
